path required argument for samba_share
Adding note in README of the required path. 

I forgot it and got an odd ValidationFailed error telling me `name is required` when it was actually the `path` I was missing. Took me a bit too long to spot this obvious error and having this line in the README would have likely shortened that debugging.